+++ poem/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- `Cors` middleware allows all HTTP methods and headers by default.
+
 # [1.0.5] 2010-10-19
 
 - Add `RouteDomain` for `Host` header routing.